# technical report/PhotoSpeak-1.0.0-APISwagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhotoSpeak",
    "description": "The API for the Photospeak - CloudComputing Project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/BADELITATIBI/PhotoSpeak/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://photospeak-project-dot-photospeak.lm.r.appspot.com/",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
    "/translate": {
      "post": {
        "tags": [
          "Translate"
        ],
        "summary": "Translate text",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/translate_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/what-is": {
      "post": {
        "tags": [
          "What-Is"
        ],
        "summary": "Get labels from a photo and translate them",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/whatis_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/getUserPhoto": {
      "get": {
        "tags": [
          "UserPhotos"
        ],
        "summary": "Get user's photos",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/speech": {
      "get": {
        "tags": [
          "TextToSpeech"
        ],
        "summary": "Convert text to speech",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "SpeechToText"
        ],
        "summary": "Check speech pronunciation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/speech_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "translate_body": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
      "whatis_body": {
        "type": "object",
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary"
          },
          "language": {
            "type": "string"
          }
        }
      },
      "speech_body": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "speech": {
            "type": "string"
          }
        }
      }
    }
  }
}